cast compute_time_ms to long long to avoid cross platform warnings when making

+++ main.cc
@@ -610,7 +610,7 @@ tool_window (mu_Context* ctx,
     mu_label (ctx, "Time Scale:");
     float_slider (ctx, time_scale);
     static char buf[64];
-    snprintf (buf, 64, "Compute Time: %ld ms", compute_time_ms);
+    snprintf (buf, 64, "Compute Time: %lld ms", (long long) compute_time_ms);
     mu_label (ctx, buf);
     mu_end_window (ctx);
   }
